fix: k,v cache use float32

+++ src/model_optimizer/models/pi05/llm.py
@@ -165,12 +165,16 @@ class LLM(torch.nn.Module, Model):
         # 导致最终 past_keys/past_values 只包含第一层。这里避免逐层索引，改为直接拼接底层缓存列表。
         if isinstance(past_key_caches, DynamicCache):
             # transformers.cache_utils.DynamicCache: key_cache/value_cache 为按层排列的 list[Tensor]
-            past_keys_tensor = torch.cat(list(past_key_caches.key_cache), dim=0).to(torch.bfloat16)
-            past_values_tensor = torch.cat(list(past_key_caches.value_cache), dim=0).to(torch.bfloat16)
+            #past_keys_tensor = torch.cat(list(past_key_caches.key_cache), dim=0).to(torch.bfloat16)
+            #past_values_tensor = torch.cat(list(past_key_caches.value_cache), dim=0).to(torch.bfloat16)
+            past_keys_tensor = torch.cat(list(past_key_caches.key_cache), dim=0)
+            past_values_tensor = torch.cat(list(past_key_caches.value_cache), dim=0)
         else:
             # 兼容 legacy cache：tuple[num_layers] of (k, v)
-            past_keys_tensor = torch.cat([kv[0] for kv in past_key_caches], dim=0).to(torch.bfloat16)
-            past_values_tensor = torch.cat([kv[1] for kv in past_key_caches], dim=0).to(torch.bfloat16)
+            #past_keys_tensor = torch.cat([kv[0] for kv in past_key_caches], dim=0).to(torch.bfloat16)
+            #past_values_tensor = torch.cat([kv[1] for kv in past_key_caches], dim=0).to(torch.bfloat16)
+            past_keys_tensor = torch.cat([kv[0] for kv in past_key_caches], dim=0)
+            past_values_tensor = torch.cat([kv[1] for kv in past_key_caches], dim=0)
 
         return past_keys_tensor, past_values_tensor, prefix_output.last_hidden_state
